# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "DryRunOut": {
        "additionalProperties": false,
        "description": "The job plan a specification renders to, with nothing allocated. AC-F14.\n\nThe point of the screen this feeds is that a specification error costs\nnothing. An allocation on this estate is the scarce resource, so the\nconsole makes the dry run the primary action and submission the secondary\none, and this response is what makes that possible: the exact command,\nenvironment and resources the scheduler would be given.",
        "properties": {
          "allocationConsumed": {
            "const": false,
            "default": false,
            "description": "Always false. A dry run that consumed an allocation would not be one.",
            "title": "Allocationconsumed",
            "type": "boolean"
          },
          "command": {
            "description": "The command, exactly as it would be run.",
            "items": {
              "type": "string"
            },
            "title": "Command",
            "type": "array"
          },
          "driver": {
            "description": "Which driver rendered the plan.",
            "title": "Driver",
            "type": "string"
          },
          "environment": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Environment the job would receive.",
            "title": "Environment",
            "type": "object"
          },
          "inputArtefactSha256": {
            "description": "The resolved digests the identity was computed over.",
            "items": {
              "type": "string"
            },
            "title": "Inputartefactsha256",
            "type": "array"
          },
          "resources": {
            "additionalProperties": true,
            "description": "What the scheduler would be asked for.",
            "title": "Resources",
            "type": "object"
          },
          "runIdentity": {
            "description": "The identity this specification would submit under (AC-F1).",
            "title": "Runidentity",
            "type": "string"
          },
          "specHash": {
            "description": "SHA-256 of the canonical specification.",
            "title": "Spechash",
            "type": "string"
          },
          "warnings": {
            "description": "Anything the driver flagged without refusing.",
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "runIdentity",
          "specHash",
          "inputArtefactSha256",
          "driver",
          "command"
        ],
        "title": "DryRunOut",
        "type": "object"
      },
...
      "Problem": {
        "description": "An RFC 9457 problem document.",
        "properties": {
          "code": {
            "description": "Machine readable, stable problem code.",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Explanation for this occurrence.",
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "URI of this occurrence.",
            "title": "Instance"
          },
          "status": {
            "description": "HTTP status code.",
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "description": "Short, human readable summary.",
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "Stable URI identifying the problem type.",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "RunSubmission": {
        "additionalProperties": false,
        "description": "A run specification being submitted. SAD 6.2.",
        "properties": {
          "specification": {
            "additionalProperties": true,
            "description": "The run specification, as SAD 6.2 defines it. Validated against the JSON Schema and hashed into the run identity.",
            "title": "Specification",
            "type": "object"
          }
        },
        "required": [
          "specification"
        ],
        "title": "RunSubmission",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nThe DRAUPNIR control plane API.\n\nErrors are RFC 9457 problem documents served as `application/problem+json`.\nMutating endpoints require an `Idempotency-Key`. Collections are cursor\npaginated. Mutable resources carry an `ETag` and require `If-Match`.\nLong operations return 202 with a run identifier; nothing blocks on training.\n",
    "summary": "CIM-56 model factory control plane",
    "title": "DRAUPNIR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events": {
      "get": {
        "description": "The run board's stream. AC-U4, AC-N3.\n\n    The per-run stream answers \"what is happening to this run\". The board needs\n    \"what is happening at this site\", and building it from one subscription per\n    visible run would open fifty-six connections to render one screen.\n\n    The connection stays open and events are pushed. A board that reconnected\n    for each answer would be polling with extra steps, which is what \"no full\n    list poll\" rules out.\n    \n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "streamSiteEvents",
        "parameters": [
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {}
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Watch this site's state deltas",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/v1/runs/dry-run": {
      "post": {
        "description": "Render the exact job plan, consuming no allocation. AC-F14.\n\n    This is the primary action of the compose screen, and submission is the\n    secondary one, because an allocation on this estate is the scarce resource\n    and a specification error should cost nothing to find.\n\n    Nothing is reserved here and nothing is recorded. `render` is pure by\n    Decision S5 and the conformance harness enforces that, so calling it is\n    the same operation whether the run is ever submitted or not -- which is\n    what makes the plan shown here the plan that would actually be run rather\n    than an approximation of it.\n    \n\nRequires: `operator`.",
        "operationId": "dryRunSpecification",
        "parameters": [
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DryRunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Render a run specification without submitting it",
        "tags": [
          "runs"
        ]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "description": "Return one run, with an `ETag` for a later conditional write.\n\nRequires: `admin`, `approver`, `curator`, `operator`, `viewer`.",
        "operationId": "getRun",
        "parameters": [
          {
            "description": "UUIDv7 run identifier.",
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "description": "UUIDv7 run identifier.",
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Correlation-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "description": "An RFC 9457 problem document.",
                  "properties": {
                    "code": {
                      "description": "Machine readable, stable problem code.",
                      "title": "Code",
                      "type": "string"
                    },
                    "detail": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "Explanation for this occurrence.",
                      "title": "Detail"
                    },
                    "instance": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "description": "URI of this occurrence.",
                      "title": "Instance"
                    },
                    "status": {
                      "description": "HTTP status code.",
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "description": "Short, human readable summary.",
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "description": "Stable URI identifying the problem type.",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "type",
                    "title",
                    "status",
                    "code"
                  ],
                  "title": "Problem",
                  "type": "object"
                }
              }
            },
            "description": "An RFC 9457 problem document"
          }
        },
        "summary": "Inspect a run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}